Clone Variable to avoid referenced value change
When the same item is present multiple times in the same sack, an
attribute will use the same Variable object (and values). The
reference to the Variable object is passed to methods that does
calculation and put the result back, leaving the new value to the
next item in the sack, making an attribute value to be calculated
multiple times and showing wrong numbers. With this change, the
Variable and it's values are cloned. Since the values are primitives,
the Array.clone() is sufficient to do the needed deep copy of the
values.