Show equipability of items as a colored background
Fix #147
(Experimental) Display a red background color on items when the selected character does not meet the requierements. 
TQVaultAE doesn't know the math requiered to compute accurately the real strength, dexterity and intelligence of a character the same way as the game does. 
It currently only use the "base" values available in the save file.